Don't nest stacks inside Drawer
Nesting the stacks allowed the drawer to be full-height, but caused
there to be multiple MapScreens generated, all of which were re-created
every time the map was shown. That's ridiculous.

With this navigation setup, the drawer is not full-height, but only one
MapScreen is generated. That tradeoff is worth it.